requirements.txt
django==5.2.9
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.1
django-celery-beat==2.6.0
reportlab==4.2.0
